requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0
gunicorn==21.2.0
